just to make sure

+++ utils/gca.c
@@ -295,19 +295,19 @@ GCAvoxelToNode(GCA *gca, MRI *mri, int xv, int yv, int zv, int *pxn,
   yscale = mri->ysize / gca->node_spacing ;
   zscale = mri->zsize / gca->node_spacing ;
 
-  *pxn = nint(xscale*(xv - mri->width/2))+gca->node_width/2;
+  *pxn = nint(xscale*((float)(xv - mri->width/2)))+gca->node_width/2;
   if (*pxn < 0)
     *pxn = 0 ;
   else if (*pxn >= gca->node_width)
     *pxn = gca->node_width-1 ;
 
-  *pyn = nint(yscale*(yv - mri->height/2))+gca->node_height/2;
+  *pyn = nint(yscale*((float)(yv - mri->height/2)))+gca->node_height/2;
   if (*pyn < 0)
     *pyn = 0 ;
   else if (*pyn >= gca->node_height)
     *pyn = gca->node_height-1 ;
 
-  *pzn = nint(zscale*(zv - mri->depth/2))+gca->node_depth/2;
+  *pzn = nint(zscale*((float)(zv - mri->depth/2)))+gca->node_depth/2;
   if (*pzn < 0)
     *pzn = 0 ;
   else if (*pzn >= gca->node_depth)
@@ -326,19 +326,19 @@ GCAvoxelToPrior(GCA *gca, MRI *mri, int xv, int yv, int zv,
   yscale = mri->ysize / gca->prior_spacing ;
   zscale = mri->zsize / gca->prior_spacing ;
 
-  *pxp = nint(xscale*(xv - mri->width/2))+gca->prior_width/2;
+  *pxp = nint(xscale*((float)(xv - mri->width/2)))+gca->prior_width/2;
   if (*pxp < 0)
     *pxp = 0 ;
   else if (*pxp >= gca->prior_width)
     *pxp = gca->prior_width-1 ;
 
-  *pyp = nint(yscale*(yv - mri->height/2))+gca->prior_height/2;
+  *pyp = nint(yscale*((float)(yv - mri->height/2)))+gca->prior_height/2;
   if (*pyp < 0)
     *pyp = 0 ;
   else if (*pyp >= gca->prior_height)
     *pyp = gca->prior_height-1 ;
 
-  *pzp = nint(zscale*(zv - mri->depth/2))+gca->prior_depth/2;
+  *pzp = nint(zscale*((float)(zv - mri->depth/2)))+gca->prior_depth/2;
   if (*pzp < 0)
     *pzp = 0 ;
   else if (*pzp >= gca->prior_depth)
@@ -357,19 +357,19 @@ GCAnodeToVoxel(GCA *gca, MRI *mri, int xn, int yn, int zn, int *pxv,
   yscale = mri->ysize / gca->node_spacing ;
   zscale = mri->zsize / gca->node_spacing ;
 
-  *pxv = nint((xn - gca->node_width/2)/xscale) + mri->width/2;
+  *pxv = nint(((float)(xn - gca->node_width/2))/xscale) + mri->width/2;
   if (*pxv < 0)
     *pxv = 0 ;
   else if (*pxv >= mri->width)
     *pxv = mri->width-1 ;
 
-  *pyv = nint((yn - gca->node_height/2)/yscale) + mri->height/2;
+  *pyv = nint(((float)(yn - gca->node_height/2))/yscale) + mri->height/2;
   if (*pyv < 0)
     *pyv = 0 ;
   else if (*pyv >= mri->height)
     *pyv = mri->height-1 ;
 
-  *pzv = nint((zn - gca->node_depth/2)/ zscale) + mri->depth/2;
+  *pzv = nint(((float)(zn - gca->node_depth/2))/ zscale) + mri->depth/2;
   if (*pzv < 0)
     *pzv = 0 ;
   else if (*pzv >= mri->depth)
@@ -387,19 +387,19 @@ GCApriorToVoxel(GCA *gca, MRI *mri, int xp, int yp, int zp, int *pxv,
   yscale = mri->ysize / gca->prior_spacing ;
   zscale = mri->zsize / gca->prior_spacing ;
 
-  *pxv = nint((xp - gca->prior_width/2)/xscale) + mri->width/2;
+  *pxv = nint(((float)(xp - gca->prior_width/2))/xscale) + mri->width/2;
   if (*pxv < 0)
     *pxv = 0 ;
   else if (*pxv >= mri->width)
     *pxv = mri->width-1 ;
 
-  *pyv = nint((yp - gca->prior_height/2)/yscale) + mri->height/2;
+  *pyv = nint(((float)(yp - gca->prior_height/2))/yscale) + mri->height/2;
   if (*pyv < 0)
     *pyv = 0 ;
   else if (*pyv >= mri->height)
     *pyv = mri->height-1 ;
 
-  *pzv = nint((zp - gca->prior_depth/2)/ zscale) + mri->depth/2;
+  *pzv = nint(((float)(zp - gca->prior_depth/2))/ zscale) + mri->depth/2;
   if (*pzv < 0)
     *pzv = 0 ;
   else if (*pzv >= mri->depth)
